implement getpass_or_cancel() without getpass(3)
get rid of a deprecated dependency which doesn't exist on android

a ^C will immediately cancel the password, without having to enter a newline.
a ^D would enter the password.
<backspace>, ^V, ^W, ^U should work as expected, but without echoing.

Instead of messing with signals, just set c_cc[VEOL] (not used by default)
to c_cc[VINTR] (^C) and disable c_cc[VINTR] -- and check the last char
in the buffer to see if the used had pressed ENTER or CTRL-C.