Also do the switch uid kludge when listening on remote forwardings
The dropbear server is always performing the tcp forwardings requested
by the user /as root/, and this has serious security implications.

The admin may have set net.ipv4.ip_unprivileged_port_start to some
other value than 1024, used 'iptables -m owner', or used some other
restriction and the running-as-root dropbear lets the user bypass it.

See also https://unix.stackexchange.com/questions/96261

Try at least to partially fix the remote fwd / listening part.

The real fix is, of course, is to switch to the logged in user as
soon as possible, but means a lot of work.

While at it, also rename pty.* to pty-util.*