dropbearkey.c Ignore unsupported command line options
To generate non interactively a key with OpenSSH the simplest command is:

ssh-keygen -t ed25519 -q -N '' -f ~/.ssh/id_ed25519

The command has two options -q quiet and -N passphrase which aren't supported by the dropbearkey.

To improve interoperability add explicit ignoring of the -q and -N with empty passphrase.
Also ignore the -v even if the DEBUG_TRACE is not set.

Signed-off-by: Sergey Ponomarev <stokito@gmail.com>