Don't try to shutdown() a pty
shutdown() of a pty doesn't work (ENOTSOCK), so we should close
it instead.

This will ensure that PTY controlling terminals are closed when a
session exits, including when multiple sessions run over a single SSH
connection.  In the normal case of a single session, the PTY controlling
terminal would be closed when the Dropbear server process exits anyway.

This possibly fixes #264 on github

It is possible that there could be subtle changes to PTY flushing
behaviour, though nothing caught by tests at present.