Add general post-quantum hybrid key exchange
This implements draft-kampanakis-curdle-ssh-pq-ke-04, a
mechanism combining a post-quantum KEM with x25519.

Adding a post-quantum KEM can avoid harvest-now-decrypt-later style
attacks (captured traffic decrypted in future by a quantum computer, if
they are created). Combining it with existing x25519 ensures that
security is not weakened compared to the present status, if a weakness
in the PQ KEM is discovered.

A future commit will add concrete PQ methods, currently this code
is unused.
x25519 could be generalised later if needed.