Fix build when htole64 macro is in (sys/)endian.h
Currently, if htole64 is available on a platform as a macro in endian.h
or sys/endian.h the AC_CHECK_DECL check does not detect them as these
files are not included in the test.  As those two headers are however
included in the build this causes the declarations of the compat
functions to be mangled by the undetected macros and fail to compile.

Resolve this by passing conditional includes of endian.h and
sys/endian.h to the includes parameter of the AC_CHECK_DECL macro so
that the htole64 macro is detected correctly.